+++ e2e-tests/development-runtime/static/used-by-off-main-thread-2.js
@@ -1,0 +1,8 @@
+{
+  const id = `script-with-src`
+  const main = document.querySelector(`main`)
+  const elem = document.createElement(`div`)
+  elem.textContent = `mutation-${id}: success`
+  elem.id = `mutation-${id}`
+  main.appendChild(elem)
+}