+++ e2e-tests/production-runtime/gatsby-config.js
@@ -26,11 +26,15 @@ module.exports = {
     },
     `gatsby-plugin-local-worker`,
     `gatsby-ssr-tsx`,
+    `gatsby-plugin-node-protocol-test`,
     `gatsby-plugin-image`,
     `gatsby-plugin-sharp`,
     `gatsby-plugin-sass`,
     `gatsby-plugin-less`,
     `gatsby-plugin-stylus`,
   ].concat(process.env.TEST_PLUGIN_OFFLINE ? [`gatsby-plugin-offline`] : []),
-  partytownProxiedURLs: [`https://unpkg.com/three@0.139.1/build/three.js`],
+  partytownProxiedURLs: [
+    `https://unpkg.com/three@0.139.1/build/three.js`,
+    `http://localhost:9000/used-by-off-main-thread-2.js`,
+  ],
 }