+++ packages/gatsby/cache-dir/head/head-export-handler-for-browser.js
@@ -9,37 +9,52 @@ import {
   headExportValidator,
   filterHeadProps,
   warnForInvalidTags,
+  diffNodes,
 } from "./utils"
 
 const hiddenRoot = document.createElement(`div`)
 
 const removePrevHeadElements = () => {
-  const prevHeadNodes = [...document.querySelectorAll(`[data-gatsby-head]`)]
-  prevHeadNodes.forEach(e => e.remove())
+  const prevHeadNodes = document.querySelectorAll(`[data-gatsby-head]`)
+
+  for (const node of prevHeadNodes) {
+    node.parentNode.removeChild(node)
+  }
 }
 
 const onHeadRendered = () => {
   const validHeadNodes = []
 
-  removePrevHeadElements()
-
   const seenIds = new Map()
   for (const node of hiddenRoot.childNodes) {
     const nodeName = node.nodeName.toLowerCase()
-    const id = node.attributes.id?.value
+    const id = node.attributes?.id?.value
 
     if (!VALID_NODE_NAMES.includes(nodeName)) {
       warnForInvalidTags(nodeName)
     } else {
-      const clonedNode = node.cloneNode(true)
+      let clonedNode = node.cloneNode(true)
       clonedNode.setAttribute(`data-gatsby-head`, true)
+
+      // Create an element for scripts to make script work
+      if (clonedNode.nodeName.toLowerCase() === `script`) {
+        const script = document.createElement(`script`)
+        for (const attr of clonedNode.attributes) {
+          script.setAttribute(attr.name, attr.value)
+        }
+        script.innerHTML = clonedNode.innerHTML
+        clonedNode = script
+      }
+
       if (id) {
         if (!seenIds.has(id)) {
           validHeadNodes.push(clonedNode)
           seenIds.set(id, validHeadNodes.length - 1)
         } else {
           const indexOfPreviouslyInsertedNode = seenIds.get(id)
-          validHeadNodes[indexOfPreviouslyInsertedNode].remove()
+          validHeadNodes[indexOfPreviouslyInsertedNode].parentNode?.removeChild(
+            validHeadNodes[indexOfPreviouslyInsertedNode]
+          )
           validHeadNodes[indexOfPreviouslyInsertedNode] = clonedNode
         }
       } else {
@@ -48,7 +63,22 @@ const onHeadRendered = () => {
     }
   }
 
-  document.head.append(...validHeadNodes)
+  const existingHeadElements = document.querySelectorAll(`[data-gatsby-head]`)
+
+  if (existingHeadElements.length === 0) {
+    document.head.append(...validHeadNodes)
+    return
+  }
+
+  const newHeadNodes = []
+  diffNodes({
+    oldNodes: existingHeadElements,
+    newNodes: validHeadNodes,
+    onStale: node => node.parentNode.removeChild(node),
+    onNew: node => newHeadNodes.push(node),
+  })
+
+  document.head.append(...newHeadNodes)
 }
 
 if (process.env.BUILD_STAGE === `develop`) {