+++ packages/gatsby/src/services/initialize.ts
@@ -497,15 +497,9 @@ export async function initialize({
   await fs.ensureDir(`${publicDirectory}/static`)
 
   // Init plugins once cache is initialized
-  if (_CFLAGS_.GATSBY_MAJOR === `4`) {
-    await apiRunnerNode(`onPluginInit`, {
-      parentSpan: activity.span,
-    })
-  } else {
-    await apiRunnerNode(`unstable_onPluginInit`, {
-      parentSpan: activity.span,
-    })
-  }
+  await apiRunnerNode(`onPluginInit`, {
+    parentSpan: activity.span,
+  })
 
   activity.end()
 
@@ -659,10 +653,12 @@ export async function initialize({
 
   const workerPool = WorkerPool.create()
 
-  // This is only run during `gatsby develop`
   if (state.config.graphqlTypegen) {
     telemetry.trackFeatureIsUsed(`GraphQLTypegen`)
-    writeGraphQLConfig(program)
+    // This is only run during `gatsby develop`
+    if (process.env.gatsby_executing_command === `develop`) {
+      writeGraphQLConfig(program)
+    }
   }
 
   return {