+++ packages/gatsby-page-utils/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.22.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-page-utils@2.22.0/packages/gatsby-page-utils) (2022-08-30)
+
+[🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.22)
+
+**Note:** Version bump only for package gatsby-page-utils
+
+## [2.21.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-page-utils@2.21.0/packages/gatsby-page-utils) (2022-08-16)
+
+[🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.21)
+
+**Note:** Version bump only for package gatsby-page-utils
+
+## [2.20.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-page-utils@2.20.0/packages/gatsby-page-utils) (2022-08-02)
+
+[🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.20)
+
+#### Chores
+
+- update dependency typescript to ^4.7.4 [#36030](https://github.com/gatsbyjs/gatsby/issues/36030) ([94a3264](https://github.com/gatsbyjs/gatsby/commit/94a32647a8c45de620d2efe99310805910586c8a))
+
 ## [2.19.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-page-utils@2.19.0/packages/gatsby-page-utils) (2022-07-19)
 
 [🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.19)